# cmd/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for Google Sheets interaction",
        "title": "Google Sheets API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Your Name",
            "url": "http://www.example.com",
            "email": "example@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/sheets/capabilities": {
            "get": {
                "description": "Reads ./capabilities.json and returns it parsed as JSON",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Utils"
                ],
                "summary": "Get capabilities",
                "responses": {
                    "200": {
                        "description": "Raw JSON",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Invalid JSON",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "404": {
                        "description": "File not found",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/sheets/clearTable": {
            "post": {
                "description": "This endpoint clear table in a Google Sheets document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "ClearTable data from Google Sheets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sheet ID",
                        "name": "sheet_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Range",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "description": "Delete request payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.DeleteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully clear table",
                        "schema": {
                            "$ref": "#/definitions/responses.ClearValuesResponse"
                        }
                    },
                    "400": {
                        "description": "Missing required parameters",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/sheets/read": {
            "get": {
                "description": "This endpoint retrieves a specified range of data from a Google Sheets document",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "Get data from Google Sheets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sheet ID",
                        "name": "sheet_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Range",
                        "name": "range",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved data",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "array",
                                "items": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Missing Sheet ID or invalid parameters",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        },
        "/api/v1/sheets/share": {
            "post": {
                "description": "Grants \"reader\" permission to a user by email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sheets"
                ],
                "summary": "Share a Google Sheet (read access)",
                "parameters": [
                    {
                        "description": "Share payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ShareRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "status + permission",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ResponseError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.DeleteRequest": {
            "type": "object",
            "properties": {
                "range": {
                    "type": "string"
                }
            }
        },
...
        "models.ShareRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "emailMessage": {
                    "type": "string"
                },
                "sendEmail": {
                    "type": "boolean"
                }
            }
        },
...
        "responses.ClearValuesResponse": {
            "type": "object",
            "properties": {
                "clearedRange": {
                    "type": "string"
                },
                "spreadsheetId": {
                    "type": "string"
                }
            }
        },
...
        "responses.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                }
            }
        },
...
    }
}